+++ SistemaNewMejia/Models/Producto.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public string NombreProducto { get; set; }
         public decimal Precio { get; set; }
-        public decimal Cantidad { get; set; }
-        public decimal Contenido { get; set; }
+        public float Cantidad { get; set; }
+        public float Contenido { get; set; }
         public int IdPresentacionProducto { get; set; }
         public PresentacionProducto PresentacionProducto { get; set; }
         public bool VenderMenudeo { get; set; }
